requirements.txt
colorama
xhtml2pdf
markdown-it-py
linkify-it-py
Pygments
